Refactor: 도메인 클래스 관련 폴더 구조 변경
- 도메인 클래스에 대하여 해당 도메인 클래스의 이름을 가지는 패키지 아래 위치하도록 변경하였습니다.
