+++ be/src/main/java/kr/codesquad/airbnb/dto/ReserveFormResponseDto.java
@@ -1,6 +1,6 @@
 package kr.codesquad.airbnb.dto;
 
-import kr.codesquad.airbnb.domain.Accommodation;
+import kr.codesquad.airbnb.domain.accommodation.Accommodation;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
