+++ be/src/main/java/kr/codesquad/airbnb/repository/MemberRepository.java
@@ -1,6 +1,6 @@
 package kr.codesquad.airbnb.repository;
 
-import kr.codesquad.airbnb.domain.Member;
+import kr.codesquad.airbnb.domain.member.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
