+++ be/src/main/java/kr/codesquad/airbnb/repository/ReservationRepository.java
@@ -1,6 +1,6 @@
 package kr.codesquad.airbnb.repository;
 
-import kr.codesquad.airbnb.domain.Reservation;
+import kr.codesquad.airbnb.domain.reservation.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
